LPD-35168 Split class into two to separate responsibilities
